refactor: 기존 HttpConstants에 있던 헤더 필드명 상수들을 유틸리티 클래스로 분리
