+++ src/main/java/webserver/utils/HttpConstants.java
@@ -1,0 +1,7 @@
+package webserver.utils;
+
+public class HttpConstants {
+    public static final String CRLF = "\r\n";
+
+    private HttpConstants() {}
+}
