+++ src/main/java/webserver/utils/HttpField.java
@@ -1,14 +1,13 @@
-package webserver.http;
+package webserver.utils;
 
-public final class HttpConstant {
+public final class HttpField {
 
     public static final String METHOD = "Method";
     public static final String URI = "URI";
     public static final String VERSION = "Version";
-    public static final String CRLF = "\r\n";
     public static final String CONTENT_TYPE = "Content-Type";
     public static final String CONTENT_LENGTH = "Content-Length";
     public static final String AUTHORIZATION = "Authorization";
 
-    private HttpConstant() {}
+    private HttpField() {}
 }
