+++ src/main/java/webserver/utils/HttpMethod.java
@@ -1,10 +1,10 @@
 package webserver.utils;
 
-public final class HttpMethodName {
+public final class HttpMethod {
     public static final String GET = "GET";
     public static final String POST = "POSt";
     public static final String PUT = "PUT";
     public static final String DELETE = "DELETE";
 
-    private HttpMethodName() {}
+    private HttpMethod() {}
 }
